+++ changelog.md
@@ -89,7 +89,12 @@ This files keeps track of all of the changes made betweem version of X Schedule.
 - Fixed bug where schedule notification trigger code would display.
 
 # 1.3.2
-- Build number: 417?
+- Build number: 428
 - Fixed major "No classes" bug where schedules would fail to load.
-- 
+ 
+# 1.3.3
+- Build number: 441
+- Fixed several bugs that caused new schedules not to display.
+  - This fixes a bug where schedules would only show the text "<p>".
+- Updated code to Swift 5.  It now compiles under Xcode 11.